add - login

+++ WebContent/day03/list.jsp
@@ -1,8 +1,29 @@
 <%@page import="java.util.List"%>
+<%@ page import="kr.co.bit.common.LoginUtil" %>
 <%@page import="kr.co.bit.dao.MemberDAO"%>
 <%@page import="kr.co.bit.vo.MemberVO"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
+<%
+	if (!LoginUtil.isLogin(request)) {
+		response.sendRedirect("../day05/loginForm.jsp");
+		return;
+	}
+	
+	boolean isMember = false;
+	MemberVO loginMember = null; 
+	String loginId = LoginUtil.getLoginId(request);
+	if (loginId.length() > 0) {
+		loginMember = (new MemberDAO()).getMember(loginId);
+		if (loginMember != null) {
+			isMember = true;
+		}
+	}
+	if (!isMember) {
+		response.sendRedirect("../day05/loginForm.jsp");
+		return;
+	}
+%>    
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
 <html>
 <head>
@@ -52,12 +73,28 @@
 		<td><%=member.getPhone() %></td>		
 		<td><%=member.getEmail() %></td>		
 		<td><%=member.getRegDate() %></td>
+<% if (member.getId().equals(loginId)) { %>
 		<td><a href="updateForm.jsp?id=<%=member.getId() %>">수정</a></td>
 		<td><a href="delete.jsp?id=<%=member.getId() %>">삭제</a></td>
+<% } else { %>
+		<td></td>
+		<td></td>
+<% } %>
 	</tr>
 <%
 	}
 %>
 </table>
+
+<%
+	if (isMember) {
+%>
+<hr>			
+아이디:<b><%=loginMember.getId() %></b>&nbsp;&nbsp;
+이 름 :<b><%=loginMember.getName() %></b>&nbsp;&nbsp;
+<input type="button" value="로그아웃" onclick="location.href='../day05/logout.jsp'">
+<%			
+	}
+%>
 </body>
 </html>
